// sm_vue_client/src/openapi/system/models/VoloAbpAccountUpdateProfileDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type VoloAbpAccountUpdateProfileDto = {
    readonly extraProperties?: Record<string, any> | null;
    userName?: string | null;
    email?: string | null;
    name?: string | null;
    surname?: string | null;
    phoneNumber?: string | null;
    concurrencyStamp?: string | null;
};